Move the jogging test into a separate file
We'll keep tests for the high-level features in separate modules now.
This is easier to spot and navigate.
